+++ lib/src/bullet_train_config.dart
@@ -1,3 +1,4 @@
+import '../bullet_train.dart';
 import 'package:dio/dio.dart';
 
 /// Default definition of connection to API
@@ -11,8 +12,8 @@ class BulletTrainConfig {
   final int receiveTimeout;
   final int sendTimeout;
 
-  final bool usePersistantStorage;
-  final String persistantDatabasePath;
+  final String storePath;
+  final StoreType storeType;
 
   /// Bullet train config initialization
   /// change only if you have self-hosted bullet train
@@ -21,10 +22,11 @@ class BulletTrainConfig {
   /// Connection settings timeouts in milliseconds
   /// [connectionTimeout], [connectTimeout], [receiveTimeout], [sendTimeout]
   ///
-  /// Persistent Storage
+  /// Storage
   ///
-  /// if you want to use Persistent storage, change [usePersistantStorage] to true
-  /// and add [persistantDatabasePath]
+  /// default type of storage used by [BulletTrainClient] is [StoreType.inMemory].
+  /// you can choose on of [StoreType.inMemory], [StoreType.prefs] and [StoreType.sembast].
+  /// for [StoreType.sembast] must by defined [storePath] where should be db file stored.
 
   const BulletTrainConfig(
       {this.baseURI = 'https://api.bullet-train.io/api/v1/',
@@ -34,8 +36,8 @@ class BulletTrainConfig {
       this.connectTimeout = 2000,
       this.receiveTimeout = 5000,
       this.sendTimeout = 5000,
-      this.usePersistantStorage = false,
-      this.persistantDatabasePath});
+      this.storeType = StoreType.inMemory,
+      this.storePath = ''});
 
   /// Client options from config
   BaseOptions get clientOptions => BaseOptions(