+++ lib/src/store/index.dart
@@ -2,3 +2,5 @@ export 'crud_store.dart';
 export 'exceptions.dart';
 export 'in_memory_store.dart';
 export 'persistant_store.dart';
+export 'persistant_sp_store.dart';
+export 'store_type.dart';