+++ lib/src/store/persistant_store.dart
@@ -84,12 +84,14 @@ class PersistantStore<T extends Flag> implements CrudStore<T> {
 
   @override
   Future<void> seed(List<T> items) async {
-    await _db.transaction((transaction) async {
-      var list = items?.map((e) => e.toMap())?.toList() ?? [];
-      if (list.isNotEmpty) {
-        await _store.addAll(transaction, list);
-      }
-    });
+    if (items != null) {
+      await _db.transaction((transaction) async {
+        var list = items?.map((e) => e.toMap())?.toList() ?? [];
+        if (list.isNotEmpty) {
+          await _store.addAll(transaction, list);
+        }
+      });
+    }
     return null;
   }
 }