+++ lib/src/store/store_type.dart
@@ -1,0 +1,2 @@
+/// [StoreType] defining type of storage used by instance of [BulletTrainClient]
+enum StoreType { inMemory, sembast, prefs }